# PICARD-1997: Reduce performance impact of fingerprint column