@phw 2020-10-28
> This is a trade between performance and fancier look, but regarding numbers it is definitively a good move.

Yes, indeed. I spent a lot of time on the fancier look and really wanted it to be like this, but with this penalty it is just not worth sticking to it. I guess some users will report the new look as an issue.